I am a fucking moron
Instead of reusing the preceding syscall instruction for a call to mmap,
we used to use the *next* syscall instruction for that. Basically,
registers were saved at the point of the *current* syscall instruction,
while memory and the rest of the process state captured the state of
execution at the *next* syscall instruction. This caused all sort of
weirdness, like code being apparently skipped for no reason, UB because
stack was overwritten, and so on.

The simplest way to reproduce the problem is to compile the following:

	#include <sys/syscall.h>

	__attribute__((noinline))
	void syscall1(long n, long a) {
	    asm volatile("syscall" : : "a"(n), "D"(a) : "rcx", "r11", "memory");
	}

	__attribute__((noinline))
	long syscall_openat(void) {
	    unsigned long ret;
	    asm volatile("syscall" : "=a"(ret) : "a"(SYS_openat), "D"(-100), "S"("/proc/self/maps"), "d"(0) : "rcx", "r11", "memory");
	    return ret;
	}

	void _start() {
	    syscall1(SYS_close, syscall_openat());
	    syscall1(SYS_exit_group, 0);
	}

...via `musl-gcc bug.c -O2 -static -nostdlib`. If the bug is triggered,
the call to close(2) should be skipped, because the return address is
overwritten.