Turn erroneously dropped controlling fd into a hard error as opposed to a hang
If, for whatever reason, a stemcell lives on without the runner being
aware of that, recvmsg will exit on EOF and the stemcell will indicate
an error code by setting state.result and triggering SIGSTOP. If this
happens in a typical runner loop, the SIGSTOP will be interpreted as the
first half of a SIGSTOP + PTRACE_EVENT_FORK indicator and thus make the
runner interpret a stemcell process as a user process. This confusion
typically leads to hangs, ILE and stuff like that. That's a bad thing;
make the bug more visible by replacing SIGSTOP with SIGKILL, which shall
trigger a hard error.