Fix hang/ECHILD on failed resume
If a resume fails, the original process could be alive or dead, in a
ptrace stop, regular stop, or any other state. Killing the cgroup brings
this down to just two cases: either the process was dead before cleanup,
or will turn dead during cleanup. The latter case is natural and is
handled correctly, but the former case (when we believe the process is
alive when it's not) will either cause ECHILD upon waitid, or cause a
hang while we're waiting for the death of the process that has already
died while we have other children.

Fix this by re-checking the health of the original process upon a
failure.