[!] Do not attempt to restore memory from a dead process
The kernel does not keep the mm pages alive if the only reference to
them is via open fds to /proc/.../mem: it requires a *process* to use
the memory as its own mm.

This means that we used to incorrectly handle shmem restoration.

Workaround this limitation by spawning another process with the same
page contents so that orig can safely die without taking its pages down.
The process is forked with CLONE_VM, so that pages are not remapped
copy-on-write, which is a slow operation.

Note that [process_isolation] may fail starting from this commit.